Fixed the bug for blocks not being deleted. I think.
I spent way too much time on this.

The most interesting thing about King Charles the First,
Is that he was 5'6" at the start of his reign,
But only 4'8" tall at the end of it.

Because of

Oliver Cromwell, Lord Protector of England (Puritan)
Born in 1599, died in 1658 (September).
Was at first (only) MP for Huntingdon (but then)
He led the Iron Side Cavalry at Marston Moor in 1644 and won.
Then he founded the New Model Army, and praise be, beat the Cavaliers at Naseby.
And the King fled up North, like an axe, to the Scots.

But under the terms of John Pimms' 'Solomn League and Covenant',
The Scots handed King Charles I over to

Oliver Cromwell, Lord Protector of England (and his warts)
Born in 1599, died in 1658 (September)
But, alas (Oi Vay!), disagreement then broke out (between)
The Presbyterian Parliament and the military
Who meant to have an independent bent and so,
The second civil war broke out and the round head ranks
Faced the cavaliers at Preston, Lancs.,
And the King lost again, silly thing (stupid King).

And Cromwell sent Colonel Pride
To purge the House of Commons of the Presbyterian Royalists,
Leaving behind only the Rump Parliament

Which appointed a High Court at Westminster Hall
To indict Charles I for Tyranny (gasp).
Charles was sentenced to death
Even though he refused to accept that the court had
Jurisdiction (say goodbye to his head).
Poor King Charles laid his head on the block (January 1649)
Down came the axe and

In the silence that followed,
The only sound that could be heard was a solitary giggle from.

Oliver Cromwell, Lord Protector of England (Ole)
Born in 1599, died in 1658 (September)
Then he smashed (Ireland), set up the Commonwealth (and more!).
He crushed the Scots at Worcester, and beat the Dutch at sea in 1653, and then

He dissolved the Rump Parliament,
And with Lambert's consent wrote the instrument of Government
Under which Oliver was Protector at last, The End!